+++ kernel/kernel.c
@@ -70,6 +70,13 @@ void kmain() {
     init_multitasking();
     init_timer(THREAD_HZ); // 100 Hz, i.e. every 10 ms
 
+    AcpiInitializeSubsystem();
+    AcpiInitializeTables(NULL, 16, FALSE);
+    AcpiLoadTables();
+
+    AcpiEnableSubsystem(ACPI_FULL_INITIALIZATION);
+    AcpiInitializeObjects(ACPI_FULL_INITIALIZATION);
+
     // Enable interrupts
     system_int_on();
 
@@ -82,12 +89,6 @@ void kmain() {
 
     create_task(shell_thread, "Shell", 0);
     create_task(handle_mouse, "Terminal mouse handler", 0);
-
-    AcpiInitializeSubsystem();
-    AcpiLoadTables();
-
-    AcpiEnableSubsystem(ACPI_FULL_INITIALIZATION);
-    AcpiInitializeObjects(ACPI_FULL_INITIALIZATION);
 
     // The OS must NEVER die.
     // Interrupts take back control from this loop whenever